expires: make expiry time accurate to nanosecond
This allows for expiry lengths to be a fraction of a second when using
the non-native expires support, and reduces expiry errors (e.g. setting
a value with an expiry at the end of a second, and having it expire too
soon).

+++ lib/moneta.rb
@@ -95,7 +95,7 @@ module Moneta
   # @api public
   def self.new(name, options = {})
     expires = options[:expires]
-    options.delete(:expires) unless Integer === expires
+    options.delete(:expires) unless Numeric === expires
     logger = options.delete(:logger)
     threadsafe = options.delete(:threadsafe)
     compress = options.delete(:compress)
